Add mpirun config.

+++ Project4/src/ising.cpp
@@ -17,7 +17,7 @@ Ising::Ising(int lattice_dimension, int rank, int world_size) {
     this->world_size = world_size;
 
     //lattice = arma::mat(lattice_dimension, lattice_dimension);
-    this->lattice = (double**) matrix(lattice_dimension, lattice_dimension, sizeof(double));
+    this->lattice = (int**) matrix(lattice_dimension, lattice_dimension, sizeof(double));
 
     double delta_energy[17];
 
@@ -38,7 +38,6 @@ void Ising::initialise_system(double temp) {
     for (int i = -8; i <= 8; i++) delta_energy[i + 8] = 0;
     for (int i = -8; i <= 8; i += 4) delta_energy[i + 8] = exp(-i/temperature);
 
-
     // Time as seed for random number generator
     srand(time(NULL));
 
@@ -46,7 +45,7 @@ void Ising::initialise_system(double temp) {
         for (int j = 0; j < lattice_dimension; j++) {
 
             // Start in random configuration
-            lattice[i][j] = (((double) rand() / (RAND_MAX)) > 0.5) ? 1: -1;
+            lattice[i][j] = ((int(rand() / (RAND_MAX)) > 0.5) ? 1: -1);
 
             // Magnetisation
             magnetisation += lattice[i][j];
@@ -118,7 +117,7 @@ void Ising::metropolis() {
 
         int delta_E = get_energy_of_site(random_x, random_y);
 
-        if (( (double) rand() * 1. / RAND_MAX) <= delta_energy[delta_E + 8]) {
+        if (( double(rand()) * 1. / RAND_MAX) <= delta_energy[delta_E + 8]) {
 
             lattice[random_x][random_y] = -1 * lattice[random_x][random_y];
             magnetisation += (double) 2 * lattice[random_x][random_y];
@@ -130,7 +129,7 @@ void Ising::metropolis() {
 }
 
 // Calculates
-double Ising::get_energy_of_site(int x, int y) {
+int Ising::get_energy_of_site(int x, int y) {
     int north = lattice[x][periodic_boundary_translation(y, lattice_dimension, -1)];
     int south = lattice[x][periodic_boundary_translation(y, lattice_dimension, 1)];
     int east  = lattice[periodic_boundary_translation(x, lattice_dimension, +1)][y];
